# SWAGGER_OPENAPI/datavault-broker-openapi.yaml
openapi: 3.1.0
info:
  title: DataVault Broker API
  description: |-
    API documentation for the DataVault Broker module.
    All endpoints require authentication via the X-UserID and X-Client-Key headers.
  version: 0.0.0 (4b97dc0)
  x-generated-at: 2026-03-12T10:56:37.585546
servers:
- url: http://localhost
  description: Generated server url
security:
- X-UserID-Header: []
- X-Client-Key-Header: []
...
paths:
...
  /vaults/{vaultId}/vaultreviews:
    get:
      tags:
      - reviews-controller
      summary: Gets a list of Reviews for a Vault
      description: Retrieves a list of all Reviews for a specific Vault.
      operationId: getVaultReviews
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      - name: vaultId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewInfo"
...
  /vaults/vaultreviews/{vaultReviewId}:
    get:
      tags:
      - reviews-controller
      summary: Get a Vault Review
      description: Retrieves a specific Vault Review by its ID.
      operationId: getVaultReview
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      - name: vaultReviewId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultReview"
...
  /vaultreviews/{vaultReviewId}/depositreviews:
    get:
      tags:
      - reviews-controller
      summary: Gets a list of Deposit Reviews for a Vault Review
      description: Retrieves a list of all Deposit Reviews for a specific Vault Review.
      operationId: getDepositReviews
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      - name: vaultReviewId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DepositReview"
  /vaultreviews/depositreviews/{depositReviewId}:
    get:
      tags:
      - reviews-controller
      summary: Gets a particular Deposit Review
      description: Retrieves a specific Deposit Review by its ID.
      operationId: getDepositReview
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      - name: depositReviewId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositReview"
  /users/{userId}:
    get:
      tags:
      - users-controller
      summary: Get a specific DataVault User
      description: Retrieves details for a specific user by their ID.
      operationId: getUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    DataCreator:
      type: object
      properties:
        id:
          type: string
        version:
          type: integer
          format: int64
        name:
          type: string
    Dataset:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        visible:
          type: boolean
        crisId:
          type: string
    RetentionPolicy:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        description:
          type: string
        url:
          type: string
        minRetentionPeriod:
          type: integer
          format: int32
        extendUponRetrieval:
          type: boolean
        inEffectDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        dataGuidanceReviewed:
          type: string
          format: date
        engine:
          type: string
          deprecated: true
        sort:
          type: integer
          format: int32
          deprecated: true
        minDataRetentionPeriod:
          type: string
          deprecated: true
        policyInfo:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        password:
          type: string
        email:
          type: string
        properties:
          type: object
          additionalProperties:
            type: string
    Vault:
      type: object
      properties:
        id:
          type: string
        version:
          type: integer
          format: int64
        creationTime:
          type: string
          format: date-time
        grantEndDate:
          type: string
          format: date
        reviewDate:
          type: string
          format: date
        name:
          type: string
        description:
          type: string
        retentionPolicy:
          $ref: "#/components/schemas/RetentionPolicy"
        retentionPolicyStatus:
          type: integer
          format: int32
        retentionPolicyExpiry:
          type: string
          format: date-time
        retentionPolicyLastChecked:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
          deprecated: true
        dataset:
          $ref: "#/components/schemas/Dataset"
        projectId:
          type: string
        snapshot:
          type: string
        affirmed:
          type: boolean
        notes:
          type: string
        estimate:
          type: string
          enum:
          - UNDER_100GB
          - UNDER_10TB
          - OVER_10TB
          - UNKNOWN
        contact:
          type: string
        pureLink:
          type: boolean
        size:
          type: integer
          format: int64
        vaultReviewUnderway:
          type: boolean
        mostRecentVaultReview:
          $ref: "#/components/schemas/VaultReview"
        dataCreator:
          type: array
          items:
            $ref: "#/components/schemas/DataCreator"
          writeOnly: true
    VaultReview:
      type: object
      properties:
        id:
          type: string
        creationTime:
          type: string
          format: date-time
        vault:
          $ref: "#/components/schemas/Vault"
        newReviewDate:
          type: string
          format: date
          deprecated: true
          writeOnly: true
        oldReviewDate:
          type: string
          format: date
        actionedDate:
          type: string
          format: date-time
        comment:
          type: string
    Deposit:
      type: object
      properties:
        id:
          type: string
          description: Universally Unique Identifier for the Deposit
        version:
          type: integer
          format: int64
        creationTime:
          type: string
          format: date-time
          description: Date that the vault was created
        vault:
          $ref: "#/components/schemas/Vault"
        depositPaths:
          type: array
          items:
            $ref: "#/components/schemas/DepositPath"
        depositChunks:
          type: array
          items:
            $ref: "#/components/schemas/DepositChunk"
        status:
          type: string
          description: Status of the Deposit
          enum:
          - NOT_STARTED
          - IN_PROGRESS
          - COMPLETE
          - FAILED
          - DELETE_IN_PROGRESS
          - DELETED
          - DELETE_FAILED
        name:
          type: string
          description: Deposit name to briefly describe the Deposit
        description:
          type: string
          description: Deposit description to provide more information about the Deposit
        hasPersonalData:
          type: boolean
          description: Whether the deposit contains personal data or not
        personalDataStatement:
          type: string
          description: Description of the nature of the personal data
        bagId:
          type: string
          description: ID of the bag associated with this Deposit
        archiveSize:
          type: integer
          format: int64
        archiveDigest:
          type: string
        archiveDigestAlgorithm:
          type: string
        numOfChunks:
          type: integer
          format: int32
        encArchiveDigest:
          type: string
        encIV:
          type: string
          format: byte
        fileOrigin:
          type: string
          description: Origin of the deposited filepath
        shortFilePath:
          type: string
          description: Short version of the origin of the deposited filepath
        filePath:
          type: string
          description: Filepath of the origin deposit
        user:
          $ref: "#/components/schemas/User"
        nonRestartJobId:
          type: string
        size:
          type: integer
          format: int64
    DepositChunk:
      type: object
      properties:
        id:
          type: string
          description: Universally Unique Identifier for the Deposit Path
        chunkNum:
          type: integer
          format: int32
          description: Filepath of the origin deposit
        archiveDigest:
          type: string
        archiveDigestAlgorithm:
          type: string
        encIV:
          type: string
          format: byte
        ecnArchiveDigest:
          type: string
    DepositPath:
      type: object
      properties:
        id:
          type: string
          description: Universally Unique Identifier for the Deposit Path
        filePath:
          type: string
          description: Filepath of the origin deposit
        pathType:
          type: string
          enum:
          - FILESTORE
          - USER_UPLOAD
    DepositReview:
      type: object
      properties:
        id:
          type: string
        creationTime:
          type: string
          format: date-time
        toBeDeleted:
          type: boolean
          deprecated: true
        deleteStatus:
          type: integer
          format: int32
        comment:
          type: string
        actionedDate:
          type: string
          format: date-time
        vaultReview:
          $ref: "#/components/schemas/VaultReview"
        deposit:
          $ref: "#/components/schemas/Deposit"
...
    ReviewInfo:
      type: object
      properties:
        vaultReviewId:
          type: string
          description: Universally Unique Identifier for the VaultReview
        depositIds:
          type: array
          description: List of all deposits associated with a VaultReview
          items:
            type: string
        depositReviewIds:
          type: array
          description: List of all depositReviews associated with a VaultReview
          items:
            type: string
...
  securitySchemes:
    X-UserID-Header:
      type: apiKey
      description: "User ID header, corresponds to `org.datavaultplatform.common.util.Constants.HEADER_USER_ID`."
      name: X-UserID
      in: header
    X-Client-Key-Header:
      type: apiKey
      description: "Client Key header, corresponds to `org.datavaultplatform.common.util.Constants.HEADER_CLIENT_KEY`."
      name: X-Client-Key
      in: header